Add Password Decryption
